+++ tsconfig.json
@@ -81,6 +81,9 @@
     "./src/Core/Materials/Base/Registerer/TextureRegister.ts",
     "./src/Core/Materials/Base/Registerer/TimeRegisterer.ts",
     "./src/Core/Materials/Base/ShaderParser.ts",
+    "./src/Core/Materials/Base/ShaderTransformers/DescriptionTransformer.ts",
+    "./src/Core/Materials/Base/ShaderTransformers/ProgramTransformer.ts",
+    "./src/Core/Materials/Base/ShaderTransformers/StringTransformer.ts",
     "./src/Core/Materials/Base/XMLRenderConfigUtility.ts",
     "./src/Core/Materials/Buffering/HitAreaMaterial.ts",
     "./src/Core/Materials/Buffering/PrimaryBufferMaterial.ts",
